Ticket 1417: HttpCookies needed to change to size_t for 64 bit.  Thanks to Bluezenix

+++ tango/net/http/HttpCookies.d
@@ -507,7 +507,7 @@ class CookieParser : Iterator!(char)
 
         ***********************************************************************/
 
-        protected uint scan (void[] data)
+        protected size_t scan (void[] data)
         {      
                 char    c;
                 int     mark,
